requirements.txt
google-generativeai
langchain
chromadb
python-dotenv
requests
beautifulsoup4
lxml
schedule